+++ packages/connect-evm/src/types.ts
@@ -12,18 +12,38 @@ export type EIP1193ProviderEvents = {
   accountsChanged: [Address[]];
   chainChanged: [Hex];
   message: [{ type: string; data: unknown }];
+  connectAndSign: [
+    { accounts: readonly Address[]; chainId: number; signResponse: string },
+  ];
+  connectWith: [
+    {
+      accounts: readonly Address[];
+      chainId: number;
+      connectWithResponse: unknown;
+    },
+  ];
 };
 
 export type EventHandlers = {
-  connect: (result: { chainId?: Hex }) => void;
+  connect: (result: { chainId: string, accounts: Address[] }) => void;
   disconnect: () => void;
   accountsChanged: (accounts: Address[]) => void;
   chainChanged: (chainId: Hex) => void;
+  connectAndSign: (result: {
+    accounts: readonly Address[];
+    chainId: number;
+    signResponse: string;
+  }) => void;
+  connectWith: (result: {
+    accounts: readonly Address[];
+    chainId: number;
+    connectWithResponse: unknown;
+  }) => void;
 };
 
 export type MetamaskConnectEVMOptions = {
   core: MultichainCore;
-  eventHandlers?: EventHandlers;
+  eventHandlers?: Partial<EventHandlers>;
   notificationQueue?: unknown[];
   supportedNetworks?: Record<CaipChainId, string>;
 };
@@ -78,7 +98,7 @@ type GenericProviderRequest = {
     | 'wallet_switchEthereumChain'
     | 'wallet_addEthereumChain'
   >;
-  params: unknown;
+  params?: unknown;
 };
 
 // Discriminated union for provider requests