+++ packages/connect-evm/CHANGELOG.md
@@ -9,8 +9,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Added `connectWith()` method which makes a request for the method to be paired together with a connection request ([#53](https://github.com/MetaMask/connect-monorepo/pull/53))
+  - This method expects param options `method`, `params`, `chainId`, and `account`
+  - `params` can be a function or a plain value. When it is a function, it will be called with the selected account as the first param and must return a value that will be used as the actual params for the request.
+- Added `eventHandlers.connectAndSign` handler which is called on successful `connectAndSign` request with an object including the `accounts`, `chainId`, and `signResponse` ([#53](https://github.com/MetaMask/connect-monorepo/pull/53))
+- Added `eventHandlers.connectWith` handler which is called on successful `connectWith` request with an object including the `accounts`, `chainId`, and `connectWithResponse` ([#53](https://github.com/MetaMask/connect-monorepo/pull/53))
 - Added analytics tracking of request handling ([#46](https://github.com/MetaMask/connect-monorepo/pull/46))
 - Initial release ([#21](https://github.com/MetaMask/connect-monorepo/pull/21))
+
+### Changed
+
+- `connect()` now always returns `chainId`, previously it could undefined ([#53](https://github.com/MetaMask/connect-monorepo/pull/53))
 
 ### Fixed